+++ pslb/pslb.config.js
@@ -60,7 +60,7 @@ module.exports = {
     repository: 'pie-framework/pie-lib',
     packages: [
       {
-        name: '@pie-lib/pie-toolbox-module',
+        name: '@pie-lib/pie-toolbox-math-rendering-module',
         output: packagesDir,
         repository: 'pie-framework/pie-lib',
         extensions: { commonJs },
@@ -75,6 +75,18 @@ module.exports = {
             'debug',
             '@pie-framework/mathquill',
           ],
+          namespace: [
+            '@pie-lib/pie-toolbox/math-rendering',
+            '@pie-lib/pie-toolbox/math-rendering-accessible',
+          ],
+        },
+      },
+      {
+        name: '@pie-lib/pie-toolbox-module',
+        output: packagesDir,
+        repository: 'pie-framework/pie-lib',
+        extensions: { commonJs },
+        imports: {
           namespace: [
             // Commented out all material-ui and d3 libs below because it looks like we don't need them for any print item
             // Also, if these are included, the build gets killed
@@ -93,7 +105,10 @@ module.exports = {
             '@pie-lib/pie-toolbox/correct-answer-toggle',
             '@pie-lib/pie-toolbox/editable-html',
             '@pie-lib/pie-toolbox/math-input',
-            '@pie-lib/pie-toolbox/math-rendering',
+            // created a separate package for math-rendering (see above)
+            '@pie-lib/pie-toolbox-math-rendering-module',
+            // '@pie-lib/pie-toolbox/math-rendering',
+            // '@pie-lib/pie-toolbox/math-rendering-accessible',
             '@pie-lib/pie-toolbox/math-toolbar',
             '@pie-lib/pie-toolbox/render-ui',
             // Not used in any print module yet: